Fix comparison between signed and unsigned integer expressions.
Only see on internal bots.

Review URL: https://codereview.chromium.org/353013008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280256 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ce0eaf006436059dfa791e4c93bcd6175a3c1aca

+++ base/android/sys_utils_unittest.cc
@@ -15,7 +15,8 @@ TEST(SysUtils, AmountOfPhysicalMemory) {
   // computed by base::SysInfo::AmountOfPhysicalMemory().
   size_t sys_ram_size =
       static_cast<size_t>(sysconf(_SC_PHYS_PAGES) * PAGE_SIZE);
-  EXPECT_EQ(sys_ram_size, SysInfo::AmountOfPhysicalMemory());
+  EXPECT_EQ(sys_ram_size,
+            static_cast<size_t>(SysInfo::AmountOfPhysicalMemory()));
 }
 
 }  // namespace android
